+++ src/main/java/de/diezwei/knotbad/knot/Operator.java
@@ -8,6 +8,4 @@ public abstract class Operator extends Knot
     public abstract AssocType getAssocType();
 
     public abstract int getPrecedence();
-    
-    public abstract String getLiteral();
 }
