+++ src/main/java/de/diezwei/knotbad/knot/Operators.java
@@ -69,4 +69,9 @@ public class Operators
     {
         return precedences.get(literal);
     }
+    
+    public Class<? extends Operator> getOperatorClass(String literal)
+    {
+        return classTypes.get(literal);
+    }
 }
